# resources/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Golf Web Specialist 1 API",
        "version": "1.0.0",
        "description": "Development-only OpenAPI documentation for auth, profile, access, security, and notification APIs."
    },
    "servers": [
        {
            "url": "http:\/\/10.16.209.234:8000\/api\/v1",
            "description": "API v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum token"
            }
        },
...
    },
    "paths": {
        "\/auth\/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create an account and issue a Sanctum token.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "name": "Jane Doe",
                                "email": "jane@example.com",
                                "password": "password",
                                "password_confirmation": "password",
                                "device_name": "iPhone 15"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Authenticated.",
                                    "data": {
                                        "user": {
                                            "id": 1,
                                            "name": "Jane Doe",
                                            "email": "jane@example.com"
                                        },
                                        "plain_text_token": "1|sanctum-token"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
...
        "\/auth\/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate with email and password.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "email": "jane@example.com",
                                "password": "password",
                                "device_name": "Postman"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated without two-factor authentication",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Authenticated.",
                                    "data": {
                                        "two_factor": false,
                                        "user": {
                                            "id": 1,
                                            "name": "Jane Doe",
                                            "email": "jane@example.com"
                                        },
                                        "plain_text_token": "1|sanctum-token"
                                    }
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Two-factor authentication is required",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Two factor authentication required.",
                                    "data": {
                                        "two_factor": true,
                                        "challenge_token": "temporary-challenge-token"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/auth\/two-factor-challenge": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Complete a two-factor login challenge with an authenticator code and issue a Sanctum token.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "challenge_token": "temporary-challenge-token",
                                "code": "123456"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated after two-factor challenge",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Authenticated.",
                                    "data": {
                                        "two_factor": true,
                                        "user": {
                                            "id": 1,
                                            "name": "Jane Doe",
                                            "email": "jane@example.com"
                                        },
                                        "plain_text_token": "1|sanctum-token"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/auth\/two-factor-recovery-code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Complete a two-factor login challenge with a recovery code and issue a Sanctum token.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "challenge_token": "temporary-challenge-token",
                                "recovery_code": "recovery-code"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated after two-factor recovery code challenge",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Authenticated.",
                                    "data": {
                                        "two_factor": true,
                                        "user": {
                                            "id": 1,
                                            "name": "Jane Doe",
                                            "email": "jane@example.com"
                                        },
                                        "plain_text_token": "1|sanctum-token"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/auth\/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current token.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
...
        "\/auth\/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send a password reset link.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "email": "jane@example.com"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/auth\/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password using an email token.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "token": "reset-token",
                                "email": "jane@example.com",
                                "password": "new-password",
                                "password_confirmation": "new-password"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
...
    }
}